Implemented first iteration for UAT
Implemented the first iteration over the unreal automation tool command
line created tool. This is not much more than a GUI wrapper to manage
the many commandline options that can be passed to the RunUAT.bat script
of UE4. It streamlines the process and allows the saving and loading of
the setup parameters.